# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api description for Hustav project",
    "version": "1.0.0",
    "title": "Hustav"
  },
  "host": "clients.onix-systems.com",
  "basePath": "/hustav-server/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/history": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Save history",
        "description": "",
        "operationId": "saveHistory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Save history",
            "required": true,
            "schema": {
              "$ref": "#/definitions/History"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SuccessData"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorWithCode"
            }
          }
        }
      },
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Get history",
        "description": "",
        "operationId": "getHistory",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": false,
            "type": "string"
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "User id",
            "required": false,
            "type": "string"
          },
          {
            "name": "action",
            "in": "query",
            "description": "Action",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "scan",
                "ratings",
                "all"
              ],
              "default": "scan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GetHistory"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorWithCode"
            }
          }
        }
      }
    },
...
    "/product": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Product",
        "description": "",
        "operationId": "product",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UcatProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ProductResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorWithCode"
            }
          }
        }
      }
    },
...
    "/manufacturer": {
      "post": {
        "tags": [
          "manufacturer"
        ],
        "summary": "Manufacturer",
        "description": "",
        "operationId": "manufacturer",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ratings List",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Manufacturer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ManufacturerDataResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorWithCode"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    }
  },
  "definitions": {
...
    "History": {
      "type": "object",
      "properties": {
        "gtin": {
          "type": "string"
        },
        "action": {
          "type": "string",
          "enum": [
            "scanned",
            "viewed",
            "voted"
          ]
        },
        "rating": {
          "type": "integer",
          "format": "int32"
        }
      },
      "xml": {
        "name": "History"
      }
    },
    "UcatProduct": {
      "type": "object",
      "properties": {
        "gtin": {
          "type": "string"
        }
      },
      "xml": {
        "name": "UcatProduct"
      }
    },
...
    "Manufacturer": {
      "type": "object",
      "properties": {
        "gln": {
          "type": "string"
        },
        "pages": {
          "type": "integer",
          "format": "int32"
        }
      },
      "xml": {
        "name": "Manufacturer"
      }
    },
...
    "ProductResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ArrayProduct"
          }
        }
      }
    },
...
    "SuccessData": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "properties": {
              "0": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "ManufacturerDataResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "properties": {
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "gln": {
                "type": "string"
              },
              "regNumber": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "nameRu": {
                "type": "string"
              },
              "nameUk": {
                "type": "string"
              },
              "nameEn": {
                "type": "string"
              },
              "shortNameRu": {
                "type": "string"
              },
              "shortNameUk": {
                "type": "string"
              },
              "shortNameEn": {
                "type": "string"
              },
              "imageLogo": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "targetMarket": {
                "type": "string"
              },
              "countPublishedProducts": {
                "type": "string"
              },
              "contacts": {
                "type": "array",
                "items": {
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "tel": {
                      "type": "string"
                    }
                  }
                }
              },
              "address": {
                "type": "string"
              },
              "Alfa2Code": {
                "type": "string"
              },
              "sort": {
                "type": "string"
              },
              "Alfa3Code": {
                "type": "string"
              },
              "LangCode": {
                "type": "string"
              },
              "codeValue": {
                "type": "string"
              },
              "use4Subscription": {
                "type": "string"
              },
              "countries_nameUk": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "GetHistory": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ArrayData"
          }
        }
      }
    },
...
    "ArrayProduct": {
      "type": "object",
      "properties": {
        "recognize_gtin": {
          "type": "integer",
          "format": "int32"
        },
        "average_rating": {
          "type": "string"
        },
        "DescriptionTextRu": {
          "type": "string"
        },
        "DescriptionTextUk": {
          "type": "string"
        },
        "DescriptionTextEn": {
          "type": "string"
        },
        "ManufacturerNameRu": {
          "type": "string"
        },
        "ManufacturerNameUk": {
          "type": "string"
        },
        "ManufacturerNameEn": {
          "type": "string"
        },
        "CountryOfOrigin": {
          "type": "string"
        },
        "FunctionalNameTextRu": {
          "type": "string"
        },
        "FunctionalNameTextUk": {
          "type": "string"
        },
        "FunctionalNameTextEn": {
          "type": "string"
        },
        "AdditionalTradeItemDescriptionRu": {
          "type": "string"
        },
        "AdditionalTradeItemDescriptionUk": {
          "type": "string"
        },
        "AdditionalTradeItemDescriptionEn": {
          "type": "string"
        },
        "UkrZed": {
          "type": "string"
        },
        "CheeseFat": {
          "type": "string"
        },
        "GeneticallyModified": {
          "type": "string"
        },
        "CaloricValue": {
          "type": "integer",
          "format": "int32"
        },
        "Fats": {
          "type": "string"
        },
        "Carbs": {
          "type": "string"
        },
        "Proteins": {
          "type": "string"
        },
        "PercentageOfAlcohol": {
          "type": "string"
        },
        "Sugar": {
          "type": "string"
        },
        "MinimumLifespanFromArrival": {
          "type": "string"
        },
        "MinimumLifespanFromProduction": {
          "type": "string"
        },
        "StorageHandlingTemperatureMaximum": {
          "type": "string"
        },
        "StorageHandlingTemperatureMinimum": {
          "type": "string"
        },
        "BrandName": {
          "type": "string"
        },
        "InformationProviderLogo": {
          "type": "string"
        },
        "Brick": {
          "type": "string"
        },
        "InformationProviderGLN": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ArrayImages"
          }
        },
        "votes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ArrayVotes"
          }
        }
      }
    },
    "ArrayData": {
      "type": "object",
      "properties": {
        "recognize_gtin": {
          "type": "integer",
          "format": "int32"
        },
        "average_rating": {
          "type": "string"
        },
        "DescriptionTextRu": {
          "type": "string"
        },
        "DescriptionTextUk": {
          "type": "string"
        },
        "DescriptionTextEn": {
          "type": "string"
        },
        "ManufacturerNameRu": {
          "type": "string"
        },
        "ManufacturerNameUk": {
          "type": "string"
        },
        "ManufacturerNameEn": {
          "type": "string"
        },
        "BrandName": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ArrayImages"
          }
        },
        "votes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ArrayVotes"
          }
        }
      }
    },
    "ArrayImages": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "link": {
          "type": "string"
        },
        "isPlanogram": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ArrayVotes": {
      "type": "object",
      "properties": {
        "fullName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "userSmallImage": {
          "type": "string"
        },
        "rating": {
          "type": "integer",
          "format": "int32"
        },
        "created": {
          "type": "string"
        }
      }
    },
...
    "ErrorWithCode": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    }
  },
...
}